@@ -9,20 +9,22 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "mpl_animators"
 description = "An interactive animation framework for matplotlib."
-requires-python = ">=3.10"
+requires-python = ">=3.12"
 readme = { file = "README.rst", content-type = "text/x-rst" }
-license = { file = "licenses/LICENSE.rst" }
+license-files = ["licenses/LICENSE.rst"]
 authors = [
   { name = "The SunPy Developers", email = "sunpy@googlegroups.com" },
 ]
 dependencies = [
-  "matplotlib>=3.5.0",
-  "numpy>=1.23.0",
-  "astropy>=5.6.0",
+  "matplotlib>=3.8.0",
+  "numpy>=1.26.0",
 ]
 dynamic = ["version"]
 
 [project.optional-dependencies]
+wcs = [
+  "astropy>=6.1.0",
+]
 tests = [
   "pytest",
   "pytest-doctestplus",
