@@ -1,3 +1,4 @@
+import warnings
 from functools import partial
 
 import numpy as np
@@ -99,8 +100,8 @@ class ArrayAnimatorWCS(ArrayAnimator):
         extra_slider_labels = []
         if "slider_functions" in kwargs and "slider_labels" not in kwargs:
             extra_slider_labels = [a.__name__ for a in kwargs['slider_functions']]
-
-        slider_labels = self._compute_slider_labels_from_wcs(slices) + extra_slider_labels
+        slider_labels = kwargs.pop('slider_labels', self._compute_slider_labels_from_wcs(slices))
+        slider_labels += extra_slider_labels
 
         super().__init__(data, image_axes=image_axes, axis_ranges=None,
                          slider_labels=slider_labels,
@@ -277,8 +278,13 @@ class ArrayAnimatorWCS(ArrayAnimator):
 
         # If we are not setting ylim globally then we set it per frame.
         if self.ylim == 'dynamic':
-            self.axes.set_ylim(float(self.data[self.frame_index].min()),
-                               float(self.data[self.frame_index].max()))
+            vmin = float(np.nanmin(self.data[self.frame_index]))
+            vmax = float(np.nanmax(self.data[self.frame_index]))
+            if np.isnan(vmin) or np.isnan(vmax):
+                warnings.warn(UserWarning(f"No data found for data slice {self.frame_index} - cannot set ylim"))
+                return
+
+            self.axes.set_ylim(vmin, vmax)
         slider.cval = val
 
     def plot_start_image_2d(self, ax):
@@ -315,7 +321,12 @@ class ArrayAnimatorWCS(ArrayAnimator):
         Get vmin, vmax of a data slice when clip_interval is specified.
         """
         percent_limits = self.clip_interval.to('%').value
-        vmin, vmax = AsymmetricPercentileInterval(*percent_limits).get_limits(self.data_transposed)
+        if np.isnan(self.data_transposed).all():
+            warnings.warn(UserWarning(f"No data found for data slice {self.frame_index} - cannot set vmin, vmax"))
+            vmin, vmax = 0, 0
+        else:
+            vmin, vmax = AsymmetricPercentileInterval(*percent_limits).get_limits(self.data_transposed)
+
         return vmin, vmax
 
     def update_plot_2d(self, val, im, slider):
